+++ src/main/java/com/ci/Cruming/CrumingApplication.java
@@ -2,8 +2,12 @@ package com.ci.Cruming;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
+@ConfigurationPropertiesScan
 @SpringBootApplication
+@EnableJpaAuditing
 public class CrumingApplication {
 
 	public static void main(String[] args) {
