+++ src/main/java/com/ci/Cruming/common/constants/Category.java
@@ -1,0 +1,6 @@
+package com.ci.Cruming.common.constants;
+
+public enum Category {
+    GENERAL,
+    PROBLEM
+}
