+++ src/main/java/com/ci/Cruming/common/constants/FileStatus.java
@@ -1,0 +1,8 @@
+package com.ci.Cruming.common.constants;
+
+import lombok.Getter;
+
+@Getter
+public enum FileStatus {
+    ACTIVE, INACTIVE
+}
