+++ src/main/java/com/ci/Cruming/common/constants/Platform.java
@@ -1,0 +1,6 @@
+package com.ci.Cruming.common.constants;
+
+public enum Platform {
+    KAKAO,
+    NAVER
+}
