+++ src/main/java/com/ci/Cruming/common/constants/UserStatus.java
@@ -1,0 +1,8 @@
+package com.ci.Cruming.common.constants;
+
+import lombok.Getter;
+
+@Getter
+public enum UserStatus {
+    ACTIVE, INACTIVE
+}
