+++ src/main/java/com/ci/Cruming/common/constants/Visibility.java
@@ -1,0 +1,7 @@
+package com.ci.Cruming.common.constants;
+
+public enum Visibility {
+    PUBLIC,
+    FOLLOWER,
+    PRIVATE
+}
