--- src/main/java/com/ci/Cruming/domain/Alarm.java
@@ -1,50 +1,0 @@
-package com.ci.Cruming.domain;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "alarm")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Alarm {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "actor_id", nullable = false)
-    private User actor;
-
-    @Column(nullable = false, length = 50)
-    private String type;
-
-    @Column(nullable = false, length = 50)
-    private String targetType;
-
-    @Column(nullable = false)
-    private Long targetId;
-
-    @Column(nullable = false, length = 450)
-    private String content;
-
-    @Column(nullable = false)
-    private boolean isRead;
-
-    @Column(nullable = false, updatable = false)
-    private LocalDateTime createdAt;
-
-    @Column
-    private LocalDateTime readAt;
-
-    @Column
-    private LocalDateTime deletedAt;
-}
