--- src/main/java/com/ci/Cruming/domain/User.java
@@ -1,69 +1,0 @@
-package com.ci.Cruming.domain;
-
-import com.ci.Cruming.domain.constants.Platform;
-import com.ci.Cruming.domain.constants.UserStatus;
-import com.ci.Cruming.domain.post.Post;
-import com.ci.Cruming.domain.timeline.Timeline;
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Table(name = "users")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, length = 50)
-    private String nickname;
-
-    @Column
-    private Short height;
-
-    @Column
-    private Short armReach;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private Platform platform;
-
-    @Column
-    private Long platformId;
-
-    @Column(length = 300)
-    private String intro;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "home_gym")
-    private Location homeGym;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private UserStatus status;
-
-    @Column(nullable = false, updatable = false)
-    private LocalDateTime createdAt;
-
-    @Column
-    private LocalDateTime deletedAt;
-
-    @OneToMany(mappedBy = "follower")
-    private List<Follow> followers = new ArrayList<>();
-
-    @OneToMany(mappedBy = "following")
-    private List<Follow> followings = new ArrayList<>();
-
-    @OneToMany(mappedBy = "user")
-    private List<Post> posts = new ArrayList<>();
-
-    @OneToMany(mappedBy = "user")
-    private List<Timeline> timelines = new ArrayList<>();
-}
