--- src/main/java/com/ci/Cruming/domain/constants/FileStatus.java
@@ -1,8 +1,0 @@
-package com.ci.Cruming.domain.constants;
-
-import lombok.Getter;
-
-@Getter
-public enum FileStatus {
-    ACTIVE, INACTIVE
-}
