--- src/main/java/com/ci/Cruming/domain/constants/Platform.java
@@ -1,6 +1,0 @@
-package com.ci.Cruming.domain.constants;
-
-public enum Platform {
-    KAKAO,
-    NAVER
-}
