--- src/main/java/com/ci/Cruming/domain/constants/Visibility.java
@@ -1,7 +1,0 @@
-package com.ci.Cruming.domain.constants;
-
-public enum Visibility {
-    PUBLIC,
-    FOLLOWER,
-    PRIVATE
-}
