--- src/main/java/com/ci/Cruming/domain/file/FileMapping.java
@@ -1,35 +1,0 @@
-package com.ci.Cruming.domain.file;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Table(name = "file_mapping")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class FileMapping {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, length = 30)
-    private String targetType;
-
-    @Column(nullable = false)
-    private Long targetId;
-
-    @Column
-    private Integer displayOrder;
-
-    @Column(nullable = false, updatable = false)
-    private LocalDateTime createdAt;
-
-    @OneToMany(mappedBy = "mapping")
-    private List<File> files = new ArrayList<>();
-}
