--- src/main/java/com/ci/Cruming/domain/post/Post.java
@@ -1,59 +1,0 @@
-package com.ci.Cruming.domain.post;
-
-import com.ci.Cruming.domain.Location;
-import com.ci.Cruming.domain.User;
-import com.ci.Cruming.domain.constants.Visibility;
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Table(name = "post")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Post {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "location_id")
-    private Location location;
-
-    @Column(length = 50)
-    private String level;
-
-    @Column(nullable = false, length = 50)
-    private String category;
-
-    @Column(nullable = false, length = 200)
-    private String title;
-
-    @Column(nullable = false, length = 3000)
-    private String content;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private Visibility visibility;
-
-    @Column(nullable = false, updatable = false)
-    private LocalDateTime createdAt;
-
-    @Column
-    private LocalDateTime deletedAt;
-
-    @OneToMany(mappedBy = "post")
-    private List<PostLike> likes = new ArrayList<>();
-
-    @OneToMany(mappedBy = "post")
-    private List<PostReply> replies = new ArrayList<>();
-}
