--- src/main/java/com/ci/Cruming/domain/post/PostLike.java
@@ -1,25 +1,0 @@
-package com.ci.Cruming.domain.post;
-
-import com.ci.Cruming.domain.User;
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "post_like")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class PostLike {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id", nullable = false)
-    private Post post;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
-}
