--- src/main/java/com/ci/Cruming/domain/timeline/TimelineReply.java
@@ -1,48 +1,0 @@
-package com.ci.Cruming.domain.timeline;
-
-import com.ci.Cruming.domain.User;
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Table(name = "timeline_reply")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class TimelineReply {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "timeline_id", nullable = false)
-    private Timeline timeline;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "parent_id")
-    private TimelineReply parent;
-
-    @OneToMany(mappedBy = "parent")
-    private List<TimelineReply> children = new ArrayList<>();
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
-
-    @Column(nullable = false, length = 1500)
-    private String content;
-
-    @Column(nullable = false, updatable = false)
-    private LocalDateTime createdAt;
-
-    @Column
-    private LocalDateTime updatedAt;
-
-    @Column
-    private LocalDateTime deletedAt;
-}
