+++ src/main/java/com/ci/Cruming/global/config/WebConfig.java
@@ -1,0 +1,23 @@
+package com.ci.Cruming.global.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebConfig implements WebMvcConfigurer {
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/**")
+                .allowedOrigins(
+                        "http://localhost:8080",
+                        "http://10.0.2.2:8080",  // 안드로이드 에뮬레이터
+                        "http://localhost:19000", // Expo 개발 서버
+                        "http://localhost:19006"  // Expo 웹
+                )
+                .allowedMethods("GET", "POST", "PUT", "DELETE", "OPTIONS")
+                .allowedHeaders("*")
+                .allowCredentials(true)
+                .maxAge(3600);
+    }
+}
