+++ src/main/java/com/ci/Cruming/location/service/LocationService.java
@@ -1,0 +1,14 @@
+package com.ci.Cruming.location.service;
+
+import com.ci.Cruming.location.repository.LocationRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional
+@RequiredArgsConstructor
+public class LocationService {
+
+    private final LocationRepository locationRepository;
+}
