+++ src/main/java/com/ci/Cruming/post/repository/PostLikeRepository.java
@@ -1,0 +1,9 @@
+package com.ci.Cruming.post.repository;
+
+import com.ci.Cruming.post.entity.PostLike;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface PostLikeRepository extends JpaRepository<PostLike, Long> {
+}
