+++ src/main/java/com/ci/Cruming/post/repository/PostReplyRepository.java
@@ -1,0 +1,10 @@
+package com.ci.Cruming.post.repository;
+
+import com.ci.Cruming.post.entity.PostReply;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface PostReplyRepository extends JpaRepository<PostReply, Long> {
+
+}
