+++ src/main/java/com/ci/Cruming/post/repository/PostRepository.java
@@ -1,0 +1,9 @@
+package com.ci.Cruming.post.repository;
+
+import com.ci.Cruming.post.entity.Post;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface PostRepository extends JpaRepository<Post, Long> {
+}
