+++ src/main/java/com/ci/Cruming/timeline/entity/TimelineLike.java
@@ -1,0 +1,25 @@
+package com.ci.Cruming.timeline.entity;
+
+import com.ci.Cruming.user.entity.User;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "timeline_like")
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class TimelineLike {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "timeline_id", nullable = false)
+    private Timeline timeline;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
+}
